+++ UnitTests/BraintreeThreeDSecureTests/BTThreeDSecureClient_Tests.swift
@@ -122,6 +122,22 @@ class BTThreeDSecureClient_Tests: XCTestCase {
         waitForExpectations(timeout: 1, handler: nil)
     }
 
+    func testPerformThreeDSecureLookup_whenCardAddChallengeRequested_sendsCardAddTrue() {
+        threeDSecureRequest.nonce = "fake-card-nonce"
+        threeDSecureRequest.amount = 9.97
+        threeDSecureRequest.dfReferenceID = "df-reference-id"
+        threeDSecureRequest.cardAddChallengeRequested = true
+
+        let expectation = expectation(description: "willCallCompletion")
+
+        client.performThreeDSecureLookup(threeDSecureRequest) { _, _ in
+            XCTAssertTrue(self.mockAPIClient.lastPOSTParameters!["cardAdd"] as! Bool)
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 1)
+    }
+
     func testPerformThreeDSecureLookup_whenSuccessful_callsBackWithResult() {
         let responseBody =
             """